refactor: 서비스 코드 인터페이스화하여 @Override로 구현
